Add header to Files area
Renames
