+++ lang/ui.es-es.json
@@ -383,6 +383,10 @@
     "defaultMessage": "Files",
     "description": "Sidebar tab label for the project files"
   },
+  "files-tab-description": {
+    "defaultMessage": "Manage your project files",
+    "description": "Subheading for the Files sidebar tab"
+  },
   "firmware-update-link": {
     "defaultMessage": "Debes <link>actualizar el firmware</link> para poder conectar a este micro:bit.",
     "description": "Text in the firmware update dialog"
