+++ lang/ui.zh-cn.json
@@ -383,6 +383,10 @@
     "defaultMessage": "Files",
     "description": "Sidebar tab label for the project files"
   },
+  "files-tab-description": {
+    "defaultMessage": "Manage your project files",
+    "description": "Subheading for the Files sidebar tab"
+  },
   "firmware-update-link": {
     "defaultMessage": "在连接 micro:bit 之前，您必须 <link>update your firmware</link>。",
     "description": "Text in the firmware update dialog"
