+++ src/messages/ui.ca.json
@@ -741,6 +741,12 @@
       "value": "Files"
     }
   ],
+  "files-tab-description": [
+    {
+      "type": 0,
+      "value": "Manage your project files"
+    }
+  ],
   "firmware-update-link": [
     {
       "type": 0,
